% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{adapt_charging_features}
\alias{adapt_charging_features}
\title{Adapt charging features}
\usage{
adapt_charging_features(
  sessions,
  time_resolution = 15,
  power_resolution = 0.01
)
}
\arguments{
\item{sessions}{tibble, sessions data set in standard format marked by \code{{evprof}} package.
The minimum required variables are:
\itemize{
\item \code{ConnectionStartDateTime} (POSIXct)
\item \code{ConnectionHours} (numeric)
\item \code{Power} (numeric)
\item \code{Energy} (numeric)
}}

\item{time_resolution}{integer, time resolution (in minutes) of the sessions' datetime variables}

\item{power_resolution}{numeric, power resolution (in kW) of the sessions' power}
}
\value{
tibble
}
\description{
Calculate connection and charging times according to energy, power and time resolution
}
\details{
All sessions' \code{Power} must be higher than \code{0}, to avoid \code{NaN} values from dividing
by zero.
The \code{ConnectionStartDateTime} is first aligned to the desired time resolution,
and the \code{ConnectionEndDateTime} is calculated according to the \code{ConnectionHours},
truncated to whole minutes. The charging time is the time needed to charge
\code{Energy} at \code{Power}, limited by the connection time. \code{Energy} is then
recalculated from \code{Power} and the charging time, rounded down to 0.01 kWh,
so it never exceeds what fits in the connection window. The
\code{ChargingEndDateTime} is the whole minute in which the energy is reached,
so the charging window always contains \code{Energy} and never ends after
\code{ConnectionEndDateTime}. \code{ConnectionHours} and \code{ChargingHours} are rounded
to 2 decimals.
}
\examples{
suppressMessages(library(dplyr))

sessions <- head(evsim::california_ev_sessions, 10)

sessions \%>\%
  select(ConnectionStartDateTime, ConnectionEndDateTime, Power)

adapt_charging_features(
  sessions,
  time_resolution = 60,
  power_resolution = 0.01
) \%>\%
  select(ConnectionStartDateTime, ConnectionEndDateTime, Power)

adapt_charging_features(
  sessions,
  time_resolution = 15,
  power_resolution = 1
) \%>\%
  select(ConnectionStartDateTime, ConnectionEndDateTime, Power)


}
