% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{add_charging_features}
\alias{add_charging_features}
\title{Change charging features with new charging power distribution}
\usage{
add_charging_features(sessions, power_rates, power_prob, resolution = 15)
}
\arguments{
\item{sessions}{tibble, sessions data set in standard format marked by \code{{evprof}} package}

\item{power_rates}{numeric vector with different charging power rates}

\item{power_prob}{numeric vector with the probability of the charging power rates with the corresponding order}

\item{resolution}{integer, time resolution (in minutes) of the sessions datetime variables}
}
\value{
tibble
}
\description{
Change charging features with new charging power distribution
}
