% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demand.R
\name{approximate_sessions}
\alias{approximate_sessions}
\title{Approximate sessions to perfect power steps}
\usage{
approximate_sessions(sessions, resolution = 15, power_interval = 0.01)
}
\arguments{
\item{sessions}{tibble, sessions data set}

\item{resolution}{integer, time resolution (in minutes) of the sessions datetime variables}

\item{power_interval}{numeric, interval of power approximation, in kW}
}
\value{
tibble
}
\description{
Approximate sessions to perfect power steps
}
