% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{california_ev_sessions}
\alias{california_ev_sessions}
\title{EV charging sessions example}
\format{
\subsection{\code{california_ev_sessions}}{

A \code{tibble} object with standard variable names defined by \code{evprof}
}
}
\source{
\url{https://ev.caltech.edu/dataset}
}
\usage{
california_ev_sessions
}
\description{
Example of an charging sessions data set ready to use by \code{evprof} functions.
It is the open source data set downloaded from the
\href{https://ev.caltech.edu/dataset}{ACN-Data website}, transformed according
to the standard names defined by \code{evprof}
}
\keyword{internal}
