% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{change_charging_features}
\alias{change_charging_features}
\title{Change charging features with new charging power distribution}
\usage{
change_charging_features(sessions, power_rates, power_prob)
}
\arguments{
\item{sessions}{tibble, sessions data set in standard format marked by \code{{evprof}} package}

\item{power_rates}{numeric vector with different charging power rates}

\item{power_prob}{numeric vector with the probability of the charging power rates with the corresponding order}
}
\value{
tibble
}
\description{
Change charging features with new charging power distribution
}
