% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{estimate_connection}
\alias{estimate_connection}
\title{Estimate sessions connection values}
\usage{
estimate_connection(n, mu, sigma, log)
}
\arguments{
\item{n}{integer, number of sessions}

\item{mu}{numeric vector, means of bivariate GMM}

\item{sigma}{numeric matrix, covariance matrix of bivariate GMM}

\item{log}{logical, true if models have logarithmic transformation}
}
\value{
vector of numeric values
}
\description{
Estimate sessions connection values following a Multi-variate Guassian
distribution.
The minimum considered value for duration is 30 minutes.
}
\keyword{internal}
