% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{estimate_energy}
\alias{estimate_energy}
\title{Estimate sessions energy values}
\usage{
estimate_energy(n, mu, sigma, log)
}
\arguments{
\item{n}{number of sessions}

\item{mu}{means of univariate GMM}

\item{sigma}{covariance matrix of univariate GMM}

\item{log}{Logical, true if models have logarithmic transformation}
}
\value{
numeric vector
}
\description{
Estimate sessions energy values
}
