% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evSessionsModule.R
\name{evSessions}
\alias{evSessions}
\title{evSessions Module: Server function}
\usage{
evSessions(
  id,
  evmodel,
  sessions_day,
  profiles_ratios,
  charging_powers,
  dates,
  resolution,
  seed = 1234
)
}
\arguments{
\item{id}{character, module ID}

\item{evmodel}{object of type \code{evmodel} (see this \href{https://mcanigueral.github.io/evprof/articles/evmodel.html}{link} for more information)}

\item{sessions_day}{tibble with variables \code{time_cycle} (names corresponding to \code{evmodel$models$time_cycle}) and \code{n_sessions} (number of daily sessions per day for each time-cycle model)}

\item{profiles_ratios}{tibble with variables \code{time_cycle}, \code{profile} and \code{ratio}
The ratios of each time_cycle must sum 100.}

\item{charging_powers}{tibble with variables \code{power} and \code{ratio}
The powers must be in kW and the ratios between 0 and 1.}

\item{dates}{reactive object returning a date sequence that will set the time frame of the simulated sessions}

\item{resolution}{reactive object returning an integer, time resolution (in minutes) of the sessions datetime variables}

\item{seed}{integer, random seed of the simulation}
}
\description{
evSessions Module: Server function
}
