% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demand.R
\name{expand_session}
\alias{expand_session}
\title{Expand a session along time slots within its connection time}
\usage{
expand_session(session, resolution)
}
\arguments{
\item{session}{tibble, sessions data set in standard format marked by \code{evprof} package
(see \href{https://mcanigueral.github.io/evprof/articles/sessions-format.html}{this article})}

\item{resolution}{integer, time resolution (in minutes) of the time slots}
}
\value{
tibble
}
\description{
The \code{session} is divided in multiple time slots
with the corresponding \code{Power} consumption, among other variables.
}
\keyword{internal}
