% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{get_charging_powers_ratios}
\alias{get_charging_powers_ratios}
\title{Get charging rates distribution in percentages}
\usage{
get_charging_powers_ratios(sessions)
}
\arguments{
\item{sessions}{sessions data set in standard format}
}
\value{
tibble
}
\description{
Get charging rates distribution in percentages
}
