% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{get_charging_rates_distribution}
\alias{get_charging_rates_distribution}
\title{Get charging rates distribution in percentages}
\usage{
get_charging_rates_distribution(sessions, unit = "year")
}
\arguments{
\item{sessions}{tibble, sessions data set in standard format marked by \code{{evprof}} package
(see \href{https://mcanigueral.github.io/evprof/articles/sessions-format.html}{this article})t}

\item{unit}{lubridate \code{floor_date} unit parameter}
}
\value{
tibble
}
\description{
Get charging rates distribution in percentages
}
