% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{get_day_sessions}
\alias{get_day_sessions}
\title{Get day sessions}
\usage{
get_day_sessions(day, ev_models, connection_log, energy_log)
}
\arguments{
\item{day}{Date to simulate}

\item{ev_models}{profiles models}

\item{connection_log}{Logical, true if connection models have logarithmic transformations}

\item{energy_log}{Logical, true if energy models have logarithmic transformations}
}
\value{
tibble
}
\description{
Get day sessions
}
