% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demand.R
\name{get_demand}
\alias{get_demand}
\title{Obtain timeseries demand from sessions dataset}
\usage{
get_demand(sessions, dttm_seq, by = "Profile")
}
\arguments{
\item{sessions}{tibble, sessions data set in standard format marked by \code{{evprof}} package}

\item{dttm_seq}{sequence of datetime values that will be the datetime variable of the returned time-series data frame}

\item{by}{character, being 'Profile' or 'Session'. When \code{by='Profile'} each column corresponds to an EV user profile.}
}
\value{
tibble
}
\description{
Obtain timeseries demand from sessions dataset
}
