% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{get_estimated_connections}
\alias{get_estimated_connections}
\title{Get estimated profiles}
\usage{
get_estimated_connections(n, profile_models, log)
}
\arguments{
\item{n}{number of sessions}

\item{profile_models}{models of the profile}

\item{log}{Logical, true if models have logarithmic transformation}
}
\value{
list with sessions connection values
}
\description{
Get estimated profiles
}
