% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evmodel.R
\name{get_evmodel_parameters}
\alias{get_evmodel_parameters}
\title{Get \code{evmodel} parameters in a list format}
\usage{
get_evmodel_parameters(evmodel)
}
\arguments{
\item{evmodel}{object of class \code{evmodel}}
}
\value{
list
}
\description{
Every time cycle is an element of the returned list, containing a list with
the user profile as elements, each one containing the ratio and the
corresponding tables with the statistic parameters of connection and
energy GMM.
}
\examples{
get_evmodel_parameters(evsim::california_ev_model)

}
