% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evmodel.R
\name{get_evmodel_summary}
\alias{get_evmodel_summary}
\title{Get \code{evmodel} parameters in a list of summary tables}
\usage{
get_evmodel_summary(evmodel)
}
\arguments{
\item{evmodel}{object of class \code{evmodel}}
}
\value{
list
}
\description{
Every time cycle is an element of the returned list, containing a table with
a user profile in every row and the mean and standard deviation values of the
GMM variables (connection duration, connection start time and energy).
If the energy models were built by charging rate, the average \code{mean} and \code{sd}
are provided without taking into account different charging rates (this
information is lost in this summary).
}
\examples{
get_evmodel_summary(evsim::california_ev_model)

}
