% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demand.R
\name{get_interval_demand}
\alias{get_interval_demand}
\title{Obtain demand from a starting dttm value and certain duration interval}
\usage{
get_interval_demand(sessions, timeslot, by = c("Profile", "Session"))
}
\arguments{
\item{sessions}{tibble, sessions data set in standard format marked by \code{{evprof}} package}

\item{timeslot}{datetime, time slot of the requested demand}

\item{by}{character, being 'Profile' or 'Session'. When \code{by='Profile'} each column corresponds to an EV user profile.}
}
\value{
tibble
}
\description{
Obtain demand from a starting dttm value and certain duration interval
}
\details{
This function is only valid if charging start/end times of sessions are aligned to a specific time-interval
}
