% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demand.R
\name{get_n_connections}
\alias{get_n_connections}
\title{Get number of existing connections}
\usage{
get_n_connections(sessions, dttm_seq)
}
\arguments{
\item{sessions}{tibble, sessions data set in standard format marked by \code{{evprof}} package}

\item{dttm_seq}{sequence of datetime values that will be the datetime variable of the returned time-series data frame}
}
\value{
tibble
}
\description{
Get number of existing connections
}
