% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demand_2.R
\name{get_n_connections_2}
\alias{get_n_connections_2}
\title{Get number of existing connections}
\usage{
get_n_connections_2(sessions, dttm_seq = NULL, by = "Profile", resolution = 15)
}
\arguments{
\item{sessions}{tibble, sessions data set in standard format marked by \code{{evprof}} package}

\item{dttm_seq}{sequence of datetime values that will be the datetime variable of the returned time-series data frame}

\item{by}{character, being 'Profile' or 'Session'. When \code{by='Profile'} each column corresponds to an EV user profile.}

\item{resolution}{integer, time resolution (in minutes) of the sessions datetime variables. If \code{dttm_seq} is defined this parameter is ignored.}
}
\value{
tibble
}
\description{
Get number of existing connections
}
