% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{get_profile_day_sessions}
\alias{get_profile_day_sessions}
\title{Get sessions for a specific day and profile}
\usage{
get_profile_day_sessions(
  profile_name,
  day,
  ev_models,
  connection_log,
  energy_log
)
}
\arguments{
\item{profile_name}{profile name}

\item{day}{day as datetime with hour 00:00}

\item{ev_models}{tibble with profiles models according to calendar}

\item{connection_log}{Logical, true if connection models have logarithmic transformations}

\item{energy_log}{Logical, true if energy models have logarithmic transformations}
}
\value{
tibble
}
\description{
Get sessions for a specific day and profile
}
