% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{get_profile_sessions}
\alias{get_profile_sessions}
\title{Get profile sessions}
\usage{
get_profile_sessions(
  profile_name,
  dates,
  ev_models,
  connection_log,
  energy_log
)
}
\arguments{
\item{profile_name}{profile name}

\item{dates}{datetime vector with dates to simualte (datetime values with hour set to 00:00)}

\item{ev_models}{profiles models}

\item{connection_log}{Logical, true if connection models have logarithmic transformations}

\item{energy_log}{Logical, true if energy models have logarithmic transformations}
}
\value{
tibble
}
\description{
Get profile sessions
}
