% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evmodel.R
\name{get_user_profiles_distribution}
\alias{get_user_profiles_distribution}
\title{Get the user profiles distribution from the original data set
used to build the model}
\usage{
get_user_profiles_distribution(evmodel)
}
\arguments{
\item{evmodel}{object of class \code{evmodel} built with \code{{evprof}}
(see this \href{https://mcanigueral.github.io/evprof/articles/evmodel.html}{link} for more information)}
}
\value{
tibble
}
\description{
Get the user profiles distribution from the original data set
used to build the model
}
