% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel-processing.R
\name{mclapply.windows}
\alias{mclapply.windows}
\title{Parallel processing for windows}
\usage{
mclapply.windows(X, FUN, ..., mc.cores)
}
\arguments{
\item{X}{a vector (atomic or list) or an expressions vector.
Other objects (including classed objects) will be coerced by as.list.}

\item{FUN}{the function to be applied to (mclapply) each element of X.}

\item{...}{For mclapply, optional named arguments to FUN.}

\item{mc.cores}{integer, number of cores to be used.
Could be overruled if number of items in list is lower.}
}
\value{
list
}
\description{
mclapply version that works on windows
}
\details{
This is a workaround to allow parallel processing in windows since
\code{parallel} package only works with linux distributions.
Source: https://www.r-bloggers.com/2014/07/implementing-mclapply-on-windows-a-primer-on-embarrassingly-parallel-computation-on-multicore-systems-with-r/
}
\keyword{internal}
