% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evmodel.R
\name{print.evmodel}
\alias{print.evmodel}
\title{\code{print} method for \code{evmodel} object class}
\usage{
\method{print}{evmodel}(obj)
}
\arguments{
\item{obj}{\code{evmodel} object}
}
\description{
\code{print} method for \code{evmodel} object class
}
