% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evmodel.R
\name{read_ev_model}
\alias{read_ev_model}
\title{Read an EV model JSON file and convert it to object of class \code{evmodel}}
\usage{
read_ev_model(file)
}
\arguments{
\item{file}{path to the JSON file}
}
\value{
object of class \code{evmodel}
}
\description{
The JSON file is generated by function
\href{https://mcanigueral.github.io/evprof/reference/save_ev_model.html}{\code{save_ev_model}}
from package \code{{evprof}}
}
