% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{round_to_interval}
\alias{round_to_interval}
\title{Round a numeric value to interval}
\usage{
round_to_interval(dbl, interval)
}
\arguments{
\item{dbl}{numeric value}

\item{interval}{decimal interval (from 0 to 1)}
}
\description{
Round a numeric value to interval
}
\keyword{internal}
