% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evmodel.R
\name{update_profiles_ratios}
\alias{update_profiles_ratios}
\title{Update the ratios of the user profiles inside \code{evmodel} object}
\usage{
update_profiles_ratios(evmodel, new_ratios, discard = FALSE)
}
\arguments{
\item{evmodel}{object of class \code{evmodel}}

\item{new_ratios}{tibble with columns: \code{time_cycle}, \code{profile}, \code{ratio}
It must have all profiles from every model, including the ones with \code{ratio = 0}.
The ratios must be between 0 and 1.}

\item{discard}{If TRUE, profiles with \code{ratio == 0} will be discarded from the \code{evmodel} object}
}
\value{
the updated \code{evmodel} object
}
\description{
Update the ratios of the user profiles inside \code{evmodel} object
}
