% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{adapt_charging_features}
\alias{adapt_charging_features}
\title{Calculate connection and charging times according to energy, power and time resolution}
\usage{
adapt_charging_features(sessions, resolution)
}
\arguments{
\item{sessions}{tibble, sessions data set in standard format marked by \code{{evprof}} package}

\item{resolution}{integer, time resolution (in minutes) of the sessions datetime variables}
}
\value{
tibble
}
\description{
The \code{ConnectionStartDateTime} is first aligned to the desired time resolution,
and the \code{ConnectionEndDateTime} is calculated according to the \code{ConnectionHours}.
The \code{ChargingHours} is recalculated with the values of \code{Energy} and \code{Power},
limited by \code{ConnectionHours}. Finally, the charging times are also calculated.
}
