% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charging_infrastructure.R
\name{add_charging_infrastructure}
\alias{add_charging_infrastructure}
\title{Assign charging station to charging sessions data set}
\usage{
add_charging_infrastructure(
  sessions,
  resolution,
  min_stations = 0,
  names_prefix = NULL,
  connections_th = 10
)
}
\arguments{
\item{sessions}{tibble, charging sessions data set}

\item{resolution}{integer, time resolution in minutes}

\item{min_stations}{integer, minimum number of charging stations to consider}

\item{names_prefix}{character, prefix of the charging station names (optional)}

\item{connections_th}{integer, minimum percentage of time intervals than a
certain number of vehicles have been connected.
This is used to avoid sizing a charging infrastructure to host for example
100 vehicles when only 5\% of time there are more than 80 vehicles connected.
Then, setting \code{connections_th = 5} will ensure that we don't over-size
the charging infrastructure. It is recommended to find this value through
multiple iterations.}
}
\value{
tibble
}
\description{
Variable \code{ChargingStation} and \code{Socket}will be assigned to the \code{sessions}
tibble with a name pattern being: \code{names_prefix} + "CHS" + number
}
