% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demand.R
\name{expand_sessions}
\alias{expand_sessions}
\title{Expand sessions along time slots}
\usage{
expand_sessions(sessions, resolution)
}
\arguments{
\item{sessions}{tibble, sessions data set in standard format marked by \code{{evprof}} package
(see \href{https://mcanigueral.github.io/evprof/articles/sessions-format.html}{this article})}

\item{resolution}{integer, time resolution (in minutes) of the time slots}
}
\value{
tibble
}
\description{
Every session in \code{sessions} is divided in multiple time slots
with the corresponding \code{Power} consumption, among other variables.
}
\keyword{internal}
