% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{get_day_sessions}
\alias{get_day_sessions}
\title{Get day sessions}
\usage{
get_day_sessions(day, ev_models, connection_log, energy_log, charging_powers)
}
\arguments{
\item{day}{Date to simulate}

\item{ev_models}{profiles models}

\item{connection_log}{Logical, true if connection models have logarithmic transformations}

\item{energy_log}{Logical, true if energy models have logarithmic transformations}

\item{charging_powers}{tibble with variables \code{power} and \code{ratio}
The powers must be in kW and the ratios between 0 and 1.}
}
\value{
tibble
}
\description{
Get day sessions
}
\keyword{internal}
