% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demand.R
\name{get_demand}
\alias{get_demand}
\title{Obtain timeseries demand from sessions dataset}
\usage{
get_demand(sessions, dttm_seq = NULL, by = "Profile", resolution = 15)
}
\arguments{
\item{sessions}{tibble, sessions data set in standard format marked by \code{{evprof}} package}

\item{dttm_seq}{sequence of datetime values that will be the datetime variable of the returned time-series data frame}

\item{by}{character, being 'Profile', 'Session' or 'ChargingStation', existing in the \code{sessions} data frame.
When \code{by='Profile'} each column corresponds to an EV user profile.}

\item{resolution}{integer, time resolution (in minutes) of the sessions datetime variables. If \code{dttm_seq} is defined this parameter is ignored.}
}
\value{
tibble
}
\description{
Obtain timeseries demand from sessions dataset
}
