% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{get_estimated_energy}
\alias{get_estimated_energy}
\title{Estimate energy given energy models tibble}
\usage{
get_estimated_energy(power_vct, energy_models, energy_log)
}
\arguments{
\item{power_vct}{numeric vector of power values of simulated sessions}

\item{energy_models}{energy models tibble}

\item{energy_log}{Logical, true if models have logarithmic transformation}
}
\value{
list of numeric vectors
}
\description{
Estimate energy given energy models tibble
}
