% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evmodel.R
\name{get_user_profiles_distribution}
\alias{get_user_profiles_distribution}
\title{User profiles distribution}
\usage{
get_user_profiles_distribution(evmodel)
}
\arguments{
\item{evmodel}{object of class \code{evmodel}}
}
\value{
tibble
}
\description{
Get the user profiles distribution from the original data set
used to build the model
}
\examples{
get_user_profiles_distribution(evsim::california_ev_model)

}
