% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evmodel.R
\name{print.evmodel}
\alias{print.evmodel}
\title{\code{print} method for EV model object of \code{evmodel} class}
\usage{
\method{print}{evmodel}(x, ...)
}
\arguments{
\item{x}{object of class \code{evmodel} built with \code{{evprof}}
(see this \href{https://mcanigueral.github.io/evprof/articles/evmodel.html}{link} for more information)}

\item{...}{further arguments passed to or from other methods.}
}
\description{
\code{print} method for EV model object of \code{evmodel} class
}
\keyword{internal}
